blog/opentofu-1-10-0-beta1: More clarity on OCI Registry Support
The original version of this showed some code examples without must context
on how to make use of them. This new version links to relevant
documentation to try to help put these examples in context, so it's
hopefully easier to figure out how and where to use them.

I also made the editorial decision to show the module call example before
the provider mirror example just because far more people write their own
modules than want to create mirrors of providers, and so it seems likely
that the module package support will be of broader interest.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>